refactor: move navigation components to header widget
- Move navigation from features/ to widgets/header/ for better FSD compliance
- Rename ClientNavigationMenu to HeaderSubmenu for clarity
- Rename ClientNavigationItem to HeaderSubmenuItem
- Update all imports and component references
- Fix CSS module imports after renaming
- Add /todo to .gitignore
