--- src/features/navigation/client/index.ts
@@ -1,1 +1,0 @@
-export {ClientNavigationMenu} from './ui/client-navigation-menu'
